# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat so to speak API",
        "description": "This is the most POWERFUL API EVER",
        "termsOfService": "",
        "contact": {
            "email": "contact@mysite.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/chat/type={type}": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "description": "Get list of chats by type",
                "operationId": "App\\Http\\Controllers\\Api\\Chat\\ChatController::index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
        "/api/chat/messages/send": {
            "post": {
                "tags": [
                    "Message"
                ],
                "summary": "Trigger method to broadcasting event of messages send",
                "operationId": "App\\Http\\Controllers\\Api\\Chat\\MessageController::sendMessage",
                "requestBody": {
                    "request": "person",
                    "description": "Optional Request Parameters for Querying",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "integer"
                            },
                            "example": "{\r\n     * chat_id=1; \r\n     * text='Some text'\r\n     * }"
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}